test(api): move sigterm smoke to binary crate
