+++ crates/au-kpis-api-http/tests/graceful_shutdown.rs
@@ -1,9 +1,4 @@
-use std::{
-    process::{Command, Stdio},
-    sync::Arc,
-    thread,
-    time::{Duration, Instant},
-};
+use std::{sync::Arc, time::Duration};
 
 use au_kpis_api_http::{AppState, router};
 use au_kpis_cache::{CacheBackend, CacheClient, CacheError, RateLimitDecision, TokenBucketConfig};
@@ -100,79 +95,4 @@ async fn server_exits_promptly_when_shutdown_token_is_cancelled() {
         .expect("server task should join");
 
     join.expect("server should exit cleanly");
-}
-
-#[test]
-fn api_binary_honors_sigterm() {
-    let manifest_dir = env!("CARGO_MANIFEST_DIR");
-    let build = Command::new("cargo")
-        .args(["build", "-p", "au-kpis-api"])
-        .current_dir(manifest_dir)
-        .status()
-        .expect("build au-kpis-api binary");
-    assert!(build.success(), "binary build failed: {build:?}");
-
-    let binary = std::path::Path::new(manifest_dir).join("../../target/debug/au-kpis-api");
-    let addr = reserve_local_addr();
-
-    let mut child = Command::new(binary)
-        .env("AU_KPIS_HTTP__BIND", &addr)
-        .env(
-            "AU_KPIS_DATABASE__URL",
-            "postgres://postgres:postgres@localhost/au_kpis",
-        )
-        .stdout(Stdio::null())
-        .stderr(Stdio::null())
-        .spawn()
-        .expect("spawn au-kpis-api");
-
-    let started = Instant::now();
-    while started.elapsed() < Duration::from_secs(10) {
-        if std::net::TcpStream::connect(&addr).is_ok() {
-            break;
-        }
-        if child
-            .try_wait()
-            .expect("poll child during startup")
-            .is_some()
-        {
-            break;
-        }
-        thread::sleep(Duration::from_millis(100));
-    }
-
-    assert!(
-        std::net::TcpStream::connect(&addr).is_ok(),
-        "au-kpis-api never became ready on {addr}"
-    );
-
-    let kill = Command::new("kill")
-        .args(["-TERM", &child.id().to_string()])
-        .status()
-        .expect("send SIGTERM");
-    assert!(kill.success(), "SIGTERM failed: {kill:?}");
-
-    let deadline = Instant::now() + Duration::from_secs(5);
-    loop {
-        if let Some(status) = child.try_wait().expect("poll child") {
-            assert!(
-                status.success(),
-                "contract server exited unsuccessfully: {status}"
-            );
-            break;
-        }
-
-        assert!(
-            Instant::now() < deadline,
-            "contract server did not exit within 5s of SIGTERM"
-        );
-        thread::sleep(Duration::from_millis(100));
-    }
-}
-
-fn reserve_local_addr() -> String {
-    let listener = std::net::TcpListener::bind("127.0.0.1:0").expect("bind ephemeral port");
-    let addr = listener.local_addr().expect("local addr");
-    drop(listener);
-    addr.to_string()
 }
